+++ pkg/providers/appsec/testdata/tcwest/modules/security/firewall.tf
@@ -1,21 +1,22 @@
-// IP/GEO Firewall
+// IP/GEO/ASN Firewall
 resource "akamai_appsec_ip_geo" "policy2" {
   config_id                  = akamai_appsec_configuration.config.config_id
   security_policy_id         = akamai_appsec_ip_geo_protection.policy2.security_policy_id
   mode                       = "block"
+  asn_network_lists          = ["119711_ASNLIST"]
   geo_network_lists          = ["32346_JTONGGEOTEST"]
   ip_network_lists           = ["16360_AISHTEST"]
   exception_ip_network_lists = ["16656_CPISERVERS"]
 }
 
-// IP/GEO Firewall
+// IP/GEO/ASN Firewall
 resource "akamai_appsec_ip_geo" "andrew" {
   config_id          = akamai_appsec_configuration.config.config_id
   security_policy_id = akamai_appsec_ip_geo_protection.andrew.security_policy_id
   mode               = "block"
 }
 
-// IP/GEO Firewall
+// IP/GEO/ASN Firewall
 resource "akamai_appsec_ip_geo" "policy1" {
   config_id                  = akamai_appsec_configuration.config.config_id
   security_policy_id         = akamai_appsec_ip_geo_protection.policy1.security_policy_id