+++ pkg/providers/cloudlets/testdata/no_activations_with_two_alb/load-balancer.tf
@@ -1,7 +1,8 @@
 resource "akamai_cloudlets_application_load_balancer" "load_balancer_test_origin" {
-  origin_id      = "test_origin"
-  description    = "test description"
-  balancing_type = "WEIGHTED"
+  origin_id          = "test_origin"
+  origin_description = ""
+  description        = "test description"
+  balancing_type     = "WEIGHTED"
 
   data_centers {
     latitude                          = 102.78108
@@ -38,9 +39,10 @@ resource "akamai_cloudlets_application_load_balancer" "load_balancer_test_origin
 }
 
 resource "akamai_cloudlets_application_load_balancer" "load_balancer_test_origin_2" {
-  origin_id      = "test_origin_2"
-  description    = "test description"
-  balancing_type = "WEIGHTED"
+  origin_id          = "test_origin_2"
+  origin_description = ""
+  description        = "test description"
+  balancing_type     = "WEIGHTED"
 
   data_centers {
     latitude                          = 102.78108