+++ pkg/providers/papi/testdata/basic-rules-datasource-serial/variables.tf
@@ -1,0 +1,19 @@
+variable "edgerc_path" {
+  type    = string
+  default = "~/.edgerc"
+}
+
+variable "config_section" {
+  type    = string
+  default = "test_section"
+}
+
+variable "contract_id" {
+  type    = string
+  default = "test_contract"
+}
+
+variable "group_id" {
+  type    = string
+  default = "grp_12345"
+}