+++ pkg/tools/flags.go
@@ -1,4 +1,4 @@
 package tools
 
-// Schema means that content of the policy will be generated using HCL instead of JSON file
-var Schema bool
+// PolicyAsHCL means that content of the policy will be generated using HCL instead of JSON file
+var PolicyAsHCL bool